# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Core Accounting MVP API
  version: 1.2.0
  description: 'مواصفات REST API للنواة المحاسبية القابلة للتوسع.


    مصادر التصميم: ERD v1.2، قاموس البيانات v1.2، مصفوفة الصلاحيات وقواعد الترحيل v1.1، وحالات الاستخدام وتدفقات العمل v1.0، مع تعديلات ADR-002 v1.3.


    قرارات تنفيذية مثبتة في هذه النسخة:

    - الشركة الحالية تُثبت في الجلسة عبر /auth/context، ولا يُعتمد companyId الوارد من العميل في العمليات التشغيلية.

    - رقم المستند يُحجز عند إنشاء المسودة مع قبول فجوات الأرقام الموثقة.

    - منشئ القيد اليدوي لا يرحله افتراضياً، ويمكن جعل ذلك إعداداً للشركات الصغيرة.

    - واجهة التخصيصات مدعومة في API، ولو أُجلت واجهتها في الإصدار الأول.

    - مدة الاحتفاظ بمفتاح Idempotency هي 24 ساعة افتراضياً وقابلة للضبط.

    - أقصى نطاق تفاعلي للتقارير 366 يوماً.


    المبالغ وأسعار الصرف والمعرفات الكبيرة تُنقل كنصوص للحفاظ على الدقة. الحقول المشفرة وBlind Indexes لا تظهر في الاستجابات
    العامة.


    تغييرات v1.1:

    - اعتماد counter_account_id رسمياً في سندات القبض والصرف العامة.

    - إضافة GET /dashboard/summary ببيانات مشتقة وصلاحيات حقلية.

    - لا تُخزن مؤشرات لوحة المتابعة كمصدر حقيقة مستقل.'
  contact:
    name: Core Accounting MVP Technical Team
  license:
    name: Internal Use
    identifier: LicenseRef-Internal
servers:
- url: /api/v1
  description: مسار نسبي يعمل على نفس أصل الواجهة في جميع البيئات
...
security:
- SessionCookie: []
paths:
...
  /reports/dashboard:
    get:
      tags:
      - Dashboard
      summary: المؤشرات المالية والتدفق النقدي للشركة الحالية
      operationId: getOperationalDashboard
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalDashboard'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: dateFrom
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        required: true
        schema:
          type: string
          format: date
      x-permission: dashboard.view
      security:
      - SessionCookie: []
  /reports/journal:
    get:
      tags:
      - Reports
      summary: تقرير دفتر اليومية
      operationId: getJournalReport
      responses:
        '200':
          description: قائمة النتائج
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JournalReportRow'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: أقصى نطاق تفاعلي 366 يوماً؛ النطاق الأكبر يحتاج مسار تصدير لاحق.
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: dateFrom
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - name: dateTo
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - name: documentType
        in: query
        schema:
          $ref: '#/components/schemas/AccountingDocumentType'
      - name: accountId
        in: query
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      x-permission: reports.journal.view
      security:
      - SessionCookie: []
  /reports/journal/csv:
    get:
      tags: [Reports]
      summary: تصدير دفتر اليومية بصيغة CSV عربية آمنة
      operationId: exportJournalReportCsv
      parameters:
      - { name: dateFrom, in: query, required: true, schema: { type: string, format: date } }
      - { name: dateTo, in: query, required: true, schema: { type: string, format: date } }
      - { name: documentType, in: query, schema: { $ref: '#/components/schemas/AccountingDocumentType' } }
      - { name: status, in: query, schema: { type: string, enum: [POSTED, REVERSED] } }
      - { name: accountId, in: query, schema: { type: string, pattern: '^[1-9][0-9]*$' } }
      - { name: search, in: query, schema: { type: string, minLength: 1, maxLength: 200 } }
      responses:
        '200': { description: 'ملف CSV حتى 10,000 قيد', content: { text/csv: { schema: { type: string } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.journal.export
      security: [{ SessionCookie: [] }]
  /reports/financial-position:
    get:
      tags: [Reports]
      summary: قائمة المركز المالي
      operationId: getFinancialPosition
      parameters:
      - { name: asOf, in: query, required: true, schema: { type: string, format: date } }
      - { name: compareAsOf, in: query, schema: { type: string, format: date } }
      - { name: includeZeroBalances, in: query, schema: { type: boolean } }
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.financial_position.view
      security: [{ SessionCookie: [] }]
  /reports/financial-position/{format}:
    get:
      tags: [Reports]
      summary: تصدير قائمة المركز المالي
      operationId: exportFinancialPosition
      parameters:
      - { name: format, in: path, required: true, schema: { type: string, enum: [csv, xlsx, pdf] } }
      - { name: asOf, in: query, required: true, schema: { type: string, format: date } }
      - { name: compareAsOf, in: query, schema: { type: string, format: date } }
      responses:
        '200': { description: ملف القائمة المالية }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.financial_statements.export
      security: [{ SessionCookie: [] }]
  /reports/income-statement:
    get:
      tags: [Reports]
      summary: قائمة الدخل مع مقارنة اختيارية
      operationId: getIncomeStatement
      parameters:
      - { name: dateFrom, in: query, required: true, schema: { type: string, format: date } }
      - { name: dateTo, in: query, required: true, schema: { type: string, format: date } }
      - { name: compareDateFrom, in: query, schema: { type: string, format: date } }
      - { name: compareDateTo, in: query, schema: { type: string, format: date } }
      - { name: includeZeroBalances, in: query, schema: { type: boolean } }
      responses:
        '200': { description: نجاح, content: { application/json: { schema: { type: object } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.income_statement.view
      security: [{ SessionCookie: [] }]
  /reports/income-statement/{format}:
    get:
      tags: [Reports]
      summary: تصدير قائمة الدخل
      operationId: exportIncomeStatement
      parameters:
      - { name: format, in: path, required: true, schema: { type: string, enum: [csv, xlsx, pdf] } }
      - { name: dateFrom, in: query, required: true, schema: { type: string, format: date } }
      - { name: dateTo, in: query, required: true, schema: { type: string, format: date } }
      responses:
        '200': { description: ملف القائمة المالية }
        '400': { $ref: '#/components/responses/ValidationError' }
      x-permission: reports.financial_statements.export
      security: [{ SessionCookie: [] }]
  /reports/ledger:
    get:
      tags:
      - Reports
      summary: تقرير الأستاذ أو كشف الطرف
      operationId: getLedgerReport
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerReport'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      description: يجب توفير واحد فقط من accountId أو customerId أو supplierId.
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: accountId
        in: query
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - name: customerId
        in: query
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - name: supplierId
        in: query
        schema:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
      - name: dateFrom
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - name: dateTo
        in: query
        required: true
        schema:
          type: string
          format: date
          example: '2026-07-30'
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      x-permission: reports.ledger.view
      security:
      - SessionCookie: []
  /reports/trial-balance:
    get:
      tags:
      - Reports
      summary: تقرير ميزان المراجعة
      operationId: getTrialBalanceReport
      responses:
        '200':
          description: نجاح
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrialBalanceReport'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessRuleViolation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - name: dateFrom
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: dateTo
        in: query
        required: true
        schema:
          type: string
          format: date
      x-permission: reports.trial_balance.view
      security:
      - SessionCookie: []
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: معرف جلسة عشوائي؛ القيمة الخام لا تُخزن في قاعدة البيانات.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: رمز Synchronizer مرتبط بالجلسة ومطلوب لكل طلب يغير الحالة.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: خطأ في التنسيق أو الحقول
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthenticated:
      description: الجلسة مفقودة أو منتهية أو ملغاة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: لا توجد صلاحية أو سياق الشركة غير مصرح به
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: الكيان غير موجود داخل الشركة الحالية؛ لا يكشف وجوده في شركة أخرى
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: تعارض إصدار أو تزامن أو تكرار
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    BusinessRuleViolation:
      description: انتهاك قاعدة أعمال، مثل فترة مغلقة أو قيد غير متوازن
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: تم تجاوز معدل الطلبات
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalError:
      description: خطأ داخلي؛ يعاد requestId فقط دون تفاصيل حساسة
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      example: '1001'
      description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
    Amount:
      type: string
      pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
      example: '1250.0000'
      description: قيمة مالية عشرية DECIMAL(19,4) كنص.
    PaginationMeta:
      type: object
      required:
      - page
      - pageSize
      - totalItems
      - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Problem:
      type: object
      required:
      - code
      - messageAr
      - requestId
      properties:
        code:
          type: string
          enum:
          - VALIDATION_ERROR
          - UNAUTHENTICATED
          - FORBIDDEN
          - NOT_FOUND
          - CONFLICT
          - BUSINESS_RULE_VIOLATION
          - RATE_LIMITED
          - INTERNAL_ERROR
        messageAr:
          type: string
        requestId:
          type: string
          format: uuid
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        details:
          type: object
          additionalProperties: true
          description: تفاصيل آمنة فقط؛ لا تتضمن SQL أو Stack Trace أو قيماً حساسة.
...
    Currency:
      type: object
      required:
      - id
      - code
      - nameAr
      - decimals
      - isActive
      properties:
        id:
          type: string
          pattern: ^[1-9][0-9]*$
        code:
          type: string
          pattern: ^[A-Z]{3}$
        nameAr:
          type: string
        decimals:
          type: integer
          minimum: 0
          maximum: 8
        isActive:
          type: boolean
...
    AccountingDocumentStatus:
      type: string
      enum:
      - DRAFT
      - POSTED
      - CANCELLED
      - REVERSED
    AccountingDocumentType:
      type: string
      enum:
      - MANUAL_JOURNAL
      - RECEIPT
      - PAYMENT
      - SALES_INVOICE
      - SALES_CREDIT_NOTE
      - PURCHASE_INVOICE
      - PURCHASE_DEBIT_NOTE
      - PERIOD_CLOSE
...
    JournalReportRow:
      type: object
      required:
      - documentId
      - documentNumber
      - documentType
      - documentDate
      - status
      - journalEntryId
      - entryNumber
      - debitTotal
      - creditTotal
      properties:
        documentId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentNumber:
          type: string
        documentType:
          $ref: '#/components/schemas/AccountingDocumentType'
        documentDate:
          type: string
          format: date
          example: '2026-07-30'
        status:
          $ref: '#/components/schemas/AccountingDocumentStatus'
        journalEntryId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        entryNumber:
          type: integer
        description:
          type: string
        debitTotal:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        creditTotal:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
    LedgerReportRow:
      type: object
      required:
      - date
      - documentNumber
      - description
      - debit
      - credit
      - runningDebit
      - runningCredit
      properties:
        date:
          type: string
          format: date
          example: '2026-07-30'
        documentId:
          type: string
          pattern: ^[1-9][0-9]*$
          example: '1001'
          description: معرّف BIGINT يُنقل كنص لتجنب فقدان الدقة في JavaScript.
        documentNumber:
          type: string
        description:
          type: string
        debit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        credit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        runningDebit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        runningCredit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
    LedgerReport:
      type: object
      required:
      - openingDebit
      - openingCredit
      - data
      - meta
      - closingDebit
      - closingCredit
      properties:
        openingDebit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        openingCredit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        data:
          type: array
          items:
            $ref: '#/components/schemas/LedgerReportRow'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        closingDebit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        closingCredit:
          type: string
          pattern: ^(0|[1-9][0-9]{0,14})\.[0-9]{4}$
          example: '1250.0000'
          description: قيمة مالية عشرية DECIMAL(19,4) كنص.
        reconciliationWarning:
          type:
          - string
          - 'null'
    TrialBalanceRow:
      type: object
      required:
      - accountId
      - code
      - nameAr
      - accountClass
      - debit
      - credit
      - balance
      properties:
        accountId:
          $ref: '#/components/schemas/Id'
        code:
          type: string
        nameAr:
          type: string
        accountClass:
          type: string
          enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
        debit:
          $ref: '#/components/schemas/Amount'
        credit:
          $ref: '#/components/schemas/Amount'
        balance:
          type: string
          pattern: ^-?(0|[1-9][0-9]{0,14})\.[0-9]{4}$
    TrialBalanceReport:
      type: object
      required:
      - range
      - data
      - totals
      properties:
        range:
          $ref: '#/components/schemas/ReportRange'
        data:
          type: array
          items:
            $ref: '#/components/schemas/TrialBalanceRow'
        totals:
          type: object
          required:
          - debit
          - credit
          properties:
            debit:
              $ref: '#/components/schemas/Amount'
            credit:
              $ref: '#/components/schemas/Amount'
    ReportRange:
      type: object
      required: [dateFrom, dateTo]
      properties:
        dateFrom: { type: string, format: date }
        dateTo: { type: string, format: date }
    SignedAmount:
      type: string
      pattern: '^-?(0|[1-9][0-9]{0,14})\.[0-9]{4}$'
    CashFlowMonth:
      type: object
      required: [month, receipts, payments, net]
      properties:
        month: { type: string, pattern: '^\d{4}-\d{2}$' }
        receipts: { $ref: '#/components/schemas/SignedAmount' }
        payments: { $ref: '#/components/schemas/SignedAmount' }
        net: { $ref: '#/components/schemas/SignedAmount' }
    OperationalDashboard:
      type: object
      required: [range, baseCurrency, metrics, cashFlow, recentActivity]
      properties:
        range: { $ref: '#/components/schemas/ReportRange' }
        baseCurrency: { $ref: '#/components/schemas/Currency' }
        metrics:
          type: object
          required: [receipts, payments, netCashFlow, activeSuppliers, activeCustomers, draftDocuments]
          properties:
            receipts: { $ref: '#/components/schemas/SignedAmount' }
            payments: { $ref: '#/components/schemas/SignedAmount' }
            netCashFlow: { $ref: '#/components/schemas/SignedAmount' }
            activeSuppliers: { type: integer, minimum: 0 }
            activeCustomers: { type: integer, minimum: 0 }
            draftDocuments: { type: integer, minimum: 0 }
        cashFlow:
          type: array
          items: { $ref: '#/components/schemas/CashFlowMonth' }
        recentActivity:
          type: array
          maxItems: 8
          items:
            type: object
            required: [id, type, documentNumber, documentDate, status, description, counterpartyName, amount]
            properties:
              id: { $ref: '#/components/schemas/Id' }
              type: { type: string, enum: [RECEIPT, PAYMENT] }
              documentNumber: { type: string }
              documentDate: { type: string, format: date }
              status: { $ref: '#/components/schemas/AccountingDocumentStatus' }
              description: { type: string }
              counterpartyName: { type: string }
              amount: { $ref: '#/components/schemas/Amount' }
...